# software_SurtiTelas.Backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SurtiTelas API",
    "version": "1.0.0",
    "description": "Backend API para la plataforma SurtiTelas - Manufactura y comercialización de prendas de vestir",
    "contact": {
      "name": "SurtiTelas Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "order-123"
          },
          "numero": {
            "type": "string",
            "example": "PED-000001"
          },
          "cliente": {
            "type": "string",
            "example": "Juan Pérez"
          },
          "asesor": {
            "type": "string",
            "example": "María Gómez"
          },
          "fecha": {
            "type": "string",
            "format": "date-time",
            "example": "2025-07-12T00:00:00.000Z"
          },
          "total": {
            "type": "number",
            "example": 50000
          },
          "items": {
            "type": "number",
            "example": 2
          },
          "estado": {
            "type": "string",
            "example": "Nuevo"
          },
          "prioridad": {
            "type": "string",
            "enum": [
              "Estándar",
              "Prioritario"
            ],
            "example": "Estándar"
          },
          "observaciones": {
            "type": "string",
            "example": "Entregar en la mañana"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-07-12T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-07-12T00:00:00.000Z"
          },
          "itemsList": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productId": {
                  "type": "string",
                  "example": "prod-123"
                },
                "nombre": {
                  "type": "string",
                  "example": "Camiseta básica"
                },
                "precio": {
                  "type": "number",
                  "example": 25000
                },
                "cantidad": {
                  "type": "number",
                  "example": 2
                }
              }
            },
            "example": [
              {
                "productId": "prod-123",
                "nombre": "Camiseta básica",
                "precio": 25000,
                "cantidad": 2
              }
            ]
          }
        }
      },
      "Paginated": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {}
          },
          "totalRecords": {
            "type": "number",
            "example": 50
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "example": 5
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "example": "cursorABC"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "not_found"
          },
          "message": {
            "type": "string",
            "example": "Recurso no encontrado"
          }
        }
      },
      "Return": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "ret-123"
          },
          "numeroDevolucion": {
            "type": "string",
            "example": "DEV-0001"
          },
          "orderId": {
            "type": "string",
            "nullable": true,
            "example": "order-123"
          },
          "prenda": {
            "type": "string",
            "nullable": true,
            "example": "Camiseta"
          },
          "referencia": {
            "type": "string",
            "nullable": true,
            "example": "REF-001"
          },
          "motivo": {
            "type": "string",
            "nullable": true,
            "example": "Talla incorrecta"
          },
          "cantidad": {
            "type": "integer",
            "example": 2
          },
          "cantidadInspeccionada": {
            "type": "integer",
            "example": 0
          },
          "fechaDevolucion": {
            "type": "string",
            "format": "date-time"
          },
          "estado": {
            "type": "string",
            "enum": [
              "RECIBIDO",
              "EN_INSPECCION",
              "APROBADO",
              "RECHAZADO",
              "EN_REPARACION",
              "REINGRESADO",
              "DESCARTADO"
            ],
            "example": "RECIBIDO"
          },
          "destino": {
            "type": "string",
            "enum": [
              "REINGRESO_INVENTARIO",
              "REPARACION",
              "DESCARTE",
              "DEVOLUCION_PROVEEDOR"
            ],
            "example": "REINGRESO_INVENTARIO"
          },
          "cliente": {
            "type": "string",
            "nullable": true,
            "example": "Juan Pérez"
          },
          "responsable": {
            "type": "string",
            "nullable": true,
            "example": "María Gómez"
          },
          "observaciones": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Domiciliario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "dom-123"
          },
          "userId": {
            "type": "string",
            "example": "user-123"
          },
          "zona": {
            "type": "string",
            "nullable": true,
            "example": "Norte"
          },
          "vehiculo": {
            "type": "string",
            "nullable": true,
            "example": "Moto"
          },
          "capacidad": {
            "type": "integer",
            "nullable": true,
            "example": 50
          },
          "activo": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CmsPage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cms-123"
          },
          "slug": {
            "type": "string",
            "example": "inicio"
          },
          "titulo": {
            "type": "string",
            "example": "Inicio"
          },
          "contenido": {
            "type": "string",
            "nullable": true,
            "example": "<p>Bienvenido</p>"
          },
          "publicado": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateDomiciliario": {
        "type": "object",
        "required": [
          "userId"
        ],
        "properties": {
          "userId": {
            "type": "string",
            "example": "user-123"
          },
          "zona": {
            "type": "string",
            "nullable": true,
            "example": "Norte"
          },
          "vehiculo": {
            "type": "string",
            "nullable": true,
            "example": "Moto"
          },
          "capacidad": {
            "type": "integer",
            "nullable": true,
            "example": 50
          }
        }
      },
...
      "CreateCmsPage": {
        "type": "object",
        "required": [
          "slug",
          "titulo"
        ],
        "properties": {
          "slug": {
            "type": "string",
            "example": "inicio"
          },
          "titulo": {
            "type": "string",
            "example": "Inicio"
          },
          "contenido": {
            "type": "string",
            "nullable": true,
            "example": "<p>Bienvenido</p>"
          },
          "publicado": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/v1/returns": {
      "get": {
        "summary": "Listar devoluciones",
        "tags": [
          "Returns"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "estado",
            "schema": {
              "type": "string",
              "enum": [
                "RECIBIDO",
                "EN_INSPECCION",
                "APROBADO",
                "RECHAZADO",
                "EN_REPARACION",
                "REINGRESADO",
                "DESCARTADO"
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de devoluciones",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Return"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Paginated"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear devolución",
        "tags": [
          "Returns"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "cantidad"
                ],
                "properties": {
                  "orderId": {
                    "type": "string"
                  },
                  "prenda": {
                    "type": "string"
                  },
                  "referencia": {
                    "type": "string"
                  },
                  "motivo": {
                    "type": "string"
                  },
                  "cantidad": {
                    "type": "integer"
                  },
                  "cantidadInspeccionada": {
                    "type": "integer"
                  },
                  "destino": {
                    "type": "string",
                    "enum": [
                      "REINGRESO_INVENTARIO",
                      "REPARACION",
                      "DESCARTE",
                      "DEVOLUCION_PROVEEDOR"
                    ]
                  },
                  "cliente": {
                    "type": "string"
                  },
                  "responsable": {
                    "type": "string"
                  },
                  "observaciones": {
                    "type": "string"
                  },
                  "fechaDevolucion": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Devolución creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Return"
                }
              }
            }
          }
        }
      }
    },
...
    "/reports/production": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Production report",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Production aggregated report"
          }
        }
      }
    },
...
    "/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "List orders",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "example": 10
            }
          },
          {
            "in": "query",
            "name": "estado",
            "schema": {
              "type": "string",
              "example": "Nuevo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated orders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedOrders"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Create order",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "clienteId": "cliente-123",
                "asesorId": "user-456",
                "prioridad": "Estándar",
                "observaciones": "Entregar en la mañana",
                "itemsList": [
                  {
                    "nombre": "Camiseta básica",
                    "precio": 25000,
                    "cantidad": 2
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Pedido creado"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "List notifications for current user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "leida",
            "schema": {
              "type": "boolean",
              "example": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "notif-123"
                          },
                          "tipo": {
                            "type": "string",
                            "example": "SUCCESS"
                          },
                          "titulo": {
                            "type": "string",
                            "example": "Pedido creado"
                          },
                          "mensaje": {
                            "type": "string",
                            "example": "Pedido PED-000001 ..."
                          },
                          "leida": {
                            "type": "boolean",
                            "example": false
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/domiciliarios": {
      "get": {
        "tags": [
          "Domiciliarios"
        ],
        "summary": "Lista domiciliarios",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "zona",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "activo",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de domiciliarios"
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          },
          "403": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
      "post": {
        "tags": [
          "Domiciliarios"
        ],
        "summary": "Crea un domiciliario",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDomiciliario"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Domiciliario creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Domiciliario"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          },
          "403": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
...
    "/cms": {
      "get": {
        "tags": [
          "CMS"
        ],
        "summary": "Lista páginas CMS",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "slug",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de páginas CMS"
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
      "post": {
        "tags": [
          "CMS"
        ],
        "summary": "Crea una página CMS",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCmsPage"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Página CMS creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CmsPage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          },
          "403": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
...
  },
...
}